Elevated: Use uint32_t instead of size_t since it differs in size

+++ games/elevated/generation/building/SplitBuildingGenerators.h
@@ -119,7 +119,7 @@ private:
 
 class ElevatorRepeater : public ElevatorGenerator {
 public:
-    ElevatorRepeater(size_t times, std::unique_ptr<ElevatorGenerator> generator)
+    ElevatorRepeater(uint32_t times, std::unique_ptr<ElevatorGenerator> generator)
         : m_times(times)
         , m_generator(std::move(generator))
     {
